utworzono "Wszystkie wydarzenia", zmieniono klase mainpage i utworzono z niej widok, aby wydarzenia mogly byc rowniez widziane we "WszystkieWydarzenia", dodano rowniez klase "BinidingClassStronaGlowna", gdzie znajduja sie potrzebne bindingi
