+++ BindingClassStronaGlowna.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Team_Manager
+{
+    public class BindingClassStronaGlowna
+    {
+        public string EventTitle{ get; set; }
+        public string EventDate { get; set; }
+        public ImageSource EventPhoto { get; set; }
+    }
+}
