+++ Team Manager.csproj
@@ -96,6 +96,9 @@
 	  <MauiXaml Update="SzablonPostWydarzenia.xaml">
 	    <Generator>MSBuild:Compile</Generator>
 	  </MauiXaml>
+	  <MauiXaml Update="WszystkieWydarzenia.xaml">
+	    <Generator>MSBuild:Compile</Generator>
+	  </MauiXaml>
 	</ItemGroup>
 
 </Project>
