+++ WszystkieWydarzenia.xaml
@@ -1,0 +1,22 @@
+<ContentPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
+             xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
+             x:Class="Team_Manager.WszystkieWydarzenia">
+
+    <ScrollView Margin="20">
+        <VerticalStackLayout Spacing="20">
+            <CollectionView ItemsSource="{Binding RecentEvents}">
+                <CollectionView.ItemTemplate>
+                    <DataTemplate>
+                        <Frame Padding="10" Margin="5">
+                            <StackLayout>
+                                <Label Text="{Binding EventTitle}" FontAttributes="Bold" />
+                                <Label Text="{Binding EventDate}" />
+                                <Image Source="{Binding EventPhoto}" HeightRequest="100" />
+                            </StackLayout>
+                        </Frame>
+                    </DataTemplate>
+                </CollectionView.ItemTemplate>
+            </CollectionView>
+        </VerticalStackLayout>
+    </ScrollView>
+</ContentPage>
